requirements.txt
# Core
fastmcp>=0.4.0
psycopg2-binary>=2.9.9
chromadb>=0.5.0

# RAG / LLM
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
sentence-transformers>=3.0.0

# PDF processing
pymupdf>=1.24.0
pdfplumber>=0.11.0

# Chat UI
streamlit>=1.35.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Policy download
datasets>=2.19.0        # HuggingFace datasets
fpdf2>=2.7.9            # Generate PDFs from text
